Loosen locks on all runtime dependencies
So we are allowed to use newer versions of those dependencies in apps that include statistics3
This was especially problematic for psych, for which version 2 would still be used throughout the whole app, while the newest version is 5.
This also made Sidekiq crash on boot when using it in combination with the statistics3 gem.